Add generateAircraft and change aircrafts

+++ src/classes/domain/aircrafts/Drones.java
@@ -2,7 +2,7 @@ package classes.domain.aircrafts;
 
 public class Drones extends Aircraft {
 
-    public Drones(String aircraftId, boolean foreign, double height, String model, int speed) {
+    public Drones(String aircraftId, boolean foreign, Integer height, String model, int speed) {
         super(aircraftId, foreign, height, model, speed);
     }
 }
