+++ src/classes/domain/aircrafts/helicopters/TransportHelicopter.java
@@ -8,23 +8,23 @@ import java.util.List;
 
 public class TransportHelicopter extends Aircraft {
 
-    private double maxCargoWeight;
+    private Integer maxCargoWeight;
 
-    public TransportHelicopter(String aircraftId, boolean foreign, double height, String model, int speed, double maxCargoWeight) {
+    public TransportHelicopter(String aircraftId, boolean foreign, Integer height, String model, int speed, Integer maxCargoWeight) {
         super(aircraftId, foreign, height, model, speed);
         this.maxCargoWeight = maxCargoWeight;
     }
 
-    public TransportHelicopter(String aircraftId, boolean foreign, double height, String model, List<Person> persons, HashMap<Integer, String> characteristics, int speed, double maxCargoWeight) {
+    public TransportHelicopter(String aircraftId, boolean foreign, Integer height, String model, List<Person> persons, HashMap<Integer, String> characteristics, int speed, Integer maxCargoWeight) {
         super(aircraftId, foreign, height, model, persons, characteristics, speed);
         this.maxCargoWeight = maxCargoWeight;
     }
 
-    public double getMaxCargoWeight() {
+    public Integer getMaxCargoWeight() {
         return maxCargoWeight;
     }
 
-    public void setMaxCargoWeight(double maxCargoWeight) {
+    public void setMaxCargoWeight(Integer maxCargoWeight) {
         this.maxCargoWeight = maxCargoWeight;
     }
